removes inspector suite builder and moves the logic to parent package
